backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
sqlalchemy==2.0.43
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
pandas==2.3.1
numpy==2.2.6
scikit-learn==1.6.1
xgboost==3.0.2
joblib==1.5.1
requests==2.32.4
python-multipart==0.0.20
python-dotenv==1.1.1
pydantic==2.11.7
schedule==1.2.0
aiohttp==3.12.15
redis==5.0.1

